dashboard/requirements.txt
streamlit
google-cloud-storage
time
datetime
pyautogui
json
streamlit_autorefresh-st_autorefresh
datetime
io
PIL-Image
numpy
matplotlib
pandas
base64
requests
